requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
commonmark==0.9.1
cryptography==3.3.1
Deprecated==1.2.10
dictdiffer==0.8.1
Flask==1.1.2
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
notion==0.0.27
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyGithub==1.54.1
PyJWT==1.7.1
pyTelegramBotAPI==3.7.4
python-dotenv==0.15.0
python-slugify==4.0.1
pytz==2020.5
requests==2.25.1
rsa==4.6
schedule==0.6.0
six==1.15.0
soupsieve==2.1
text-unidecode==1.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
